Give UI to the game
I added an event listener and buttons that allow
a player to play the computer by pushing these
buttons.

I also added a style sheet to make some things nicer.

The logic inside the event listener plays until either
player or computer reach 5 points. It continually
shows and updates the scoreboard and evaluates the
game in the end.